+++ README.md
@@ -20,7 +20,7 @@
   - [x] 구매할 프로모션 상품 만큼 재고 수량 차감하는 기능
   - [x] 구매할 상품개수가 재고에 충분한지 확인하는 기능
   - [x] 구매할 일반 상품 재고가 있는지 확인하는 기능
-  - [ ] 구매할 프로모션 상품 재고가 있는지 확인하는 기능
+  - [x] 구매할 프로모션 상품 재고가 있는지 확인하는 기능
   - [x] 상품을 구매하면 재고를 차감한다.
 
 ## 📝프로모션 할인
@@ -28,7 +28,7 @@
   - [ ] 예외 상황
     - [ ] 프로모션 정보형식이 유효하지 않으면 `[ERROR] 프로모션 형식이 유효하지 않습니다.`
   - [x] 프로모션 상품인지 확인하는 기능
-  - [ ] 프로모션 구매 수량만큼 재고창고에서 차감하는 기능추가
+  - [x] 프로모션 구매 수량만큼 재고창고에서 차감하는 기능추가
   - [ ] 프로모션 기간 확인하는 기능
   - [ ] 프로모션 개수 buy, get 알려주는 기능
   - [ ] 프로모션 개수만큼 가져왔는지 확인하는 기능
