+++ src/main/java/store/model/ShoppingItem.java
@@ -2,7 +2,7 @@ package store.model;
 
 public class ShoppingItem {
     private final String name;
-    private final int quantity;
+    private int quantity;
 
     public ShoppingItem(String name, int quantity) {
         this.name = name;
@@ -15,5 +15,9 @@ public class ShoppingItem {
 
     public int getQuantity() {
         return quantity;
+    }
+
+    public void decreaseQuantityBy(int notApplyPromotionQuantity) {
+        this.quantity -= notApplyPromotionQuantity;
     }
 }
