+++ src/main/java/store/view/OutputView.java
@@ -43,4 +43,9 @@ public class OutputView {
     public void display(String message) {
         System.out.println(message);
     }
+
+    public void displayConfirmPurchase(String productName, int notApplyPromotionQuantity) {
+        System.out.printf("현재 %s %d개는 프로모션 할인이 적용되지 않습니다. 그래도 구매하시겠습니까? (Y/N)",
+                productName, notApplyPromotionQuantity);
+    }
 }
